Rename doc2vec to doc2vec vectorizer
