+++ wellcomeml/ml/__init__.py
@@ -14,7 +14,7 @@ if development_transformers_mode:
     from .bert_semantic_equivalence import SemanticEquivalenceClassifier
 else:
     from .frequency_vectorizer import WellcomeTfidf
-    from .doc2vec import Doc2VecVectorizer
+    from .doc2vec_vectorizer import Doc2VecVectorizer
 
     try:
         from .vectorizer import Vectorizer
