+++ wellcomeml/ml/doc2vec_vectorizer.py
